Fixes for prometheus-related tests.

+++ cmd/runner/limiter.go
@@ -55,7 +55,7 @@ func limitCheck(acts *activity) (limit bool, msg string) {
 	// See if the total of running tasks and ran tasks equals or exceed the maximum number
 	// this runner has been configured to handle
 	//if *maxTasksOpt != 0 && math.Round(running+ran) > math.Round(float64(*maxTasksOpt)) {
-	if *maxTasksOpt != 0 && ran >= *maxTasksOpt {
+	if *maxTasksOpt != 0 && ran >= (int32)(*maxTasksOpt) {
 		// See if we are drained and the max run count has been reached
 		if running != 0 {
 			msg = fmt.Sprint("stack", stack.Trace().TrimRuntime())
